requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
google-generativeai==0.3.1
python-dotenv==1.0.0
SQLAlchemy==2.0.25
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
dateparser==1.1.8
scikit-learn==1.3.0
numpy==1.24.4
pandas==2.0.3
markdown==3.5.1 